# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "GeneratedAppOut": {
        "properties": {
          "app_type": {
            "default": "dashboard",
            "title": "App Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Release Id"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "releases": {
            "items": {
              "$ref": "#/components/schemas/AppReleaseOut"
            },
            "title": "Releases",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visibility": {
            "enum": [
              "private",
              "public"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "visibility",
          "current_release_id",
          "releases",
          "created_at",
          "updated_at"
        ],
        "title": "GeneratedAppOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Durable chat, cited local knowledge, provenance, and permissioned read-only tools.",
    "title": "Agentic Knowledge Workspace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/apps": {
      "get": {
        "operationId": "list_apps_api_apps_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeneratedAppOut"
                  },
                  "title": "Response List Apps Api Apps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Apps",
        "tags": [
          "generated-apps"
        ]
      },
      "post": {
        "operationId": "create_app_api_apps_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratedAppCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedAppOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create App",
        "tags": [
          "generated-apps"
        ]
      }
    },
    "/api/apps/{app_id}/generate": {
      "post": {
        "operationId": "generate_app_release_api_apps__app_id__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedAppOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate App Release",
        "tags": [
          "generated-apps"
        ]
      }
    },
    "/api/apps/{app_id}/preview": {
      "get": {
        "operationId": "preview_app_api_apps__app_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppPreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview App",
        "tags": [
          "generated-apps"
        ]
      }
    },
    "/api/apps/{app_id}/releases": {
      "post": {
        "operationId": "create_release_api_apps__app_id__releases_post",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppReleaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedAppOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Release",
        "tags": [
          "generated-apps"
        ]
      }
    },
    "/api/apps/{app_id}/releases/{release_id}/frame": {
      "get": {
        "operationId": "app_release_frame_api_apps__app_id__releases__release_id__frame_get",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "App Release Frame",
        "tags": [
          "generated-apps"
        ]
      }
    },
    "/api/apps/{app_id}/releases/{release_id}/publish": {
      "post": {
        "operationId": "publish_release_api_apps__app_id__releases__release_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "release_id",
            "required": true,
            "schema": {
              "title": "Release Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedAppOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Release",
        "tags": [
          "generated-apps"
        ]
      }
    },
...
    "/published/apps/{slug}": {
      "get": {
        "operationId": "published_app_published_apps__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedAppOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Published App",
        "tags": [
          "generated-apps"
        ]
      }
    },
...
  }
}